add LineBuffer to UdpWriter and UnixWriter
This change adds a LineBuffer, which can be optionally enabled for the
UdpWriter and the UnixWriter. It is disabled by default, to ensure that
the standard operation of this client works in as many contexts as
possible.

The LineBuffer is intended to help increase the performance of delivering
protocol lines to the spectatord process, when many metrics are created by
an application.

Testing buffering with the UdpWriter and UnixWriter, we can increase the
performance of these writers from ~80K rps and ~90K rps to ~150K rps, using
a 2KB buffer size. A 1KB buffer size is a little slower, and a 4KB buffer
size does not offer more than a little extra performance. With buffering
enabled, we do not see much of a difference between UDP and UDS.

The LineBuffer includes a background thread, which will flush metrics every
five seconds. This covers the case where metrics are sent at a low volume,
while ensuring timely delivery, but it also means that care should be taken
when initializing the Registry with this option, particularly in pre-fork
environments such as gunicorn.

The Config gains a new parameter, which is used to set a buffer_size, and it
defaults to zero.

Logging for Registry and Writer initialization has been adjusted. The logs
for the GlobalRegistry are squelched, to prevent spamming logs, when the
library is simply imported, but custom Registry instantiations will provide
some info level logs. Since the Meters are exported at the top-level of the
library, the writer parameter is now an Optional defaulting to None, to avoid
spamming logs with NoopWriter instantiations. The standard usage patterns for
this library will not allow an AttributeError to occur from poorly constructed
Meters.